test(learnings): isolate packed budget checker
